Reset edit state before sorting
This change improved the architecture a little bit. Now there is more
control over editing and edit state is not bound to an individual cell.
Instead that information is retained outside of the cells. It is better
this way because cells can get sorted.

Previously as cells contained edit state, they could remain editable
even after sorting.