sort - Sort based on resolved values
Now `sorter` handles resolving. Performance-wise this isn't ideal. I
thought about resolving on a higher level. The problem is that certain
operations, like edit, rely on the original value. Therefore we would
need to split the data somehow and that would break a lot of the current
API contracts.

Perhaps some kind of a resolution pass is still the way to go, though.
Just need to find the right design for it.

Closes #160.