highlighting - Rewrite so that it works with the new optimizations
Still need to test this against a nested property and document.

Instead of `highlight('foo')` we have a `highlight` pass that has to be
applied on the data after searching. That will generate metadata for the
`highlighted` formatter.

The implementation isn't ideal as we have to duplicate the search result
structure. But that is needed to avoid mutation and associated problems.