docs - Restore selection example
This has to go through data manipulation for the optimizations to work.
Earlier it relied on manipulating state and performing a check on
`render`. Now we skip a lot of that work and refresh only the rows that
changed (i.e. previously selected row and the newly selected one) so the
implementation has to be a little different as well.