feat(easy) - Push `onMoveColumns` out of table core
This is a breaking change that pushes full control over drag and drop to
the user. As a side benefit the table doesn't have to maintain column
state anymore (single source of truth -> no sync bugs) but you do have
to implement `onMoveColumns` prop for this to work. It's a `setState` at
minimum.