refactor: Support the same protocol for each renderer
This can be improved further. Now rows can do a bit more while the rest
render just functions. It probably should be the same for each. Then you
could use factories and so on.

It may also make sense to drop `Provider` and pass renderers explicitly
per component as that feels cleaner.